+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Raise an informative error if the parameters include a key that does not exist in the `conditions` section.
+- A page control can be hidden when the value of the key parameter in the URL matches one of the list of `hide_for_key` values in the list definition.
 ### Changed
 ### Fixed
 - Exception was raised when the params included a key for a multiselect and there were no conditions with the same key. This was because the multiselect key was being used as the conditions key.